(+) expose all valid LED modes from the system76 keyboard firmware
